Removed AUC as a separate tool, modified old validation tool implementations
