+++ eis_toolkit/validation/plot_prediction_area_curves.py
@@ -2,11 +2,23 @@ import matplotlib
 import numpy as np
 import pandas as pd
 from beartype import beartype
-from beartype.typing import Union
+from beartype.typing import Tuple, Union
 from matplotlib import pyplot as plt
+from shapely.geometry import LineString
 
 from eis_toolkit.exceptions import InvalidParameterValueException
-from eis_toolkit.validation.get_pa_intersection import get_pa_intersection
+
+
+def _get_pa_intersection(
+    true_positive_rate_values: Union[np.ndarray, pd.Series],
+    proportion_of_area_values: Union[np.ndarray, pd.Series],
+    threshold_values: Union[np.ndarray, pd.Series],
+) -> Tuple[float, float]:
+    true_positive_area_curve = LineString(np.column_stack((threshold_values, true_positive_rate_values)))
+    proportion_of_area_values_curve = LineString(np.column_stack((threshold_values, 1 - proportion_of_area_values)))
+    intersection = true_positive_area_curve.intersection(proportion_of_area_values_curve)
+
+    return intersection.x, intersection.y
 
 
 def _plot_prediction_area_curves(
@@ -14,7 +26,7 @@ def _plot_prediction_area_curves(
     proportion_of_area_values: Union[np.ndarray, pd.Series],
     threshold_values: Union[np.ndarray, pd.Series],
 ) -> matplotlib.figure.Figure:
-    intersection = get_pa_intersection(true_positive_rate_values, proportion_of_area_values, threshold_values)
+    x, y = _get_pa_intersection(true_positive_rate_values, proportion_of_area_values, threshold_values)
 
     fig = plt.figure(figsize=(10, 7))
     ax1 = fig.add_subplot(111)
@@ -22,7 +34,7 @@ def _plot_prediction_area_curves(
     ax1.plot(threshold_values, true_positive_rate_values, "r-", label="Prediction rate")
 
     ax2.plot(threshold_values, proportion_of_area_values, "b-", label="Area")
-    ax2.plot(intersection[0], 1 - intersection[1], " o", markersize=7, c="black", label="Intersection point")
+    ax2.plot(x, 1 - y, " o", markersize=7, c="black", label="Intersection point")
     ax1.set_ylim(0, 1.01)
     ax2.set_ylim(-0.01, 1)
     ax2.invert_yaxis()
@@ -30,9 +42,9 @@ def _plot_prediction_area_curves(
     ax1.set_ylabel("True positive rate", color="r")
     ax2.set_ylabel("Proportion of area", color="b")
     ax1.annotate(
-        text="TPR:" + str(round(intersection[1], 2)),
-        xy=(intersection[0], intersection[1]),
-        xytext=(intersection[0] + threshold_values.max() / 10, intersection[1]),
+        text="TPR:" + str(round(y, 2)),
+        xy=(x, y),
+        xytext=(x + threshold_values.max() / 10, y),
         arrowprops=dict(facecolor="black", shrink=0.09, width=0.3),
         verticalalignment="center",
     )
