+++ eis_toolkit/validation/plot_rate_curve.py
@@ -2,11 +2,11 @@ import matplotlib
 import numpy as np
 import pandas as pd
 from beartype import beartype
-from beartype.typing import Literal, Union
+from beartype.typing import Union
 from matplotlib import pyplot as plt
+from sklearn.metrics import auc
 
 from eis_toolkit.exceptions import InvalidParameterValueException
-from eis_toolkit.validation.calculate_auc import calculate_auc
 
 
 def _plot_rate_curve(
@@ -20,8 +20,8 @@ def _plot_rate_curve(
     plt.ylabel("True positive rate")
     plt.plot([0, 1], [0, 1], "--", label="Random baseline")
     auc_bbox = dict(boxstyle="round", facecolor="grey", alpha=0.2)
-    auc = str(round(calculate_auc(x_values, y_values), 2))
-    plt.text(0.8, 0.2, "AUC: " + auc, bbox=auc_bbox)
+    auc_value = str(round(auc(x_values, y_values), 2))
+    plt.text(0.8, 0.2, "AUC: " + auc_value, bbox=auc_bbox)
     plt.title(label)
     fig.legend(bbox_to_anchor=(0.85, 0.4))
 
@@ -32,36 +32,26 @@ def _plot_rate_curve(
 def plot_rate_curve(
     x_values: Union[np.ndarray, pd.Series],
     y_values: Union[np.ndarray, pd.Series],
-    plot_type: Literal["success_rate", "prediction_rate", "roc"] = "success_rate",
+    plot_title: str = "success_rate",
 ) -> matplotlib.figure.Figure:
-    """Plot success rate, prediction rate or ROC curve.
+    """Plot success rate.
 
-    Plot type depends on plot_type argument. Y-axis is always true positive rate, while x-axis can be either false
-    positive rate (roc) or proportion of area (success and prediction rate) depending on plot type.
+    Y-axis is true positive rate and x-axis is proportion of area.
 
     Args:
-        x_values: False positive rate values or proportion of area values.
+        x_values: Proportion of area values.
         y_values: True positive rate values.
-        plot_type: Plot type. Can be either: "success_rate", "prediction_rate" or "roc".
+        plot_title: Success rate
 
     Returns:
-        Success rate, prediction rate or ROC plot figure object.
+        Matplotlib figure containing the produced plot.
 
     Raises:
         InvalidParameterValueException: Invalid plot type.
         InvalidParameterValueException: x_values or y_values are out of bounds.
     """
-    if plot_type == "success_rate":
-        label = "Success rate"
-        xlab = "Proportion of area"
-    elif plot_type == "prediction_rate":
-        label = "Prediction rate"
-        xlab = "Proportion of area"
-    elif plot_type == "roc":
-        label = "ROC"
-        xlab = "False positive rate"
-    else:
-        raise InvalidParameterValueException("Invalid plot type")
+    label = plot_title
+    xlab = "Proportion of area"
 
     if x_values.max() > 1 or x_values.min() < 0:
         raise InvalidParameterValueException("x_values should be within range 0-1")
